requirements.txt
python-telegram-bot==20.7
pandas==2.2.2
matplotlib==3.8.2
arabic-reshaper==3.0.0
python-bidi==0.4.2
numpy==1.26.2
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.9